admin/Backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-jwt-extended
flask-migrate
marshmallow
chart.js
react-chartjs-2
flask-jwt-extended
passlib
bcrypt